+++ src/main/java/com/tencentcloudapi/as/v20180419/models/DataDisk.java
@@ -36,7 +36,7 @@ public class DataDisk extends AbstractModel{
     */
     @SerializedName("DiskSize")
     @Expose
-    private Long DiskSize;
+    private Integer DiskSize;
 
     /**
     * 数据盘快照 ID，类似 `snap-l8psqwnt`。
@@ -72,7 +72,7 @@ public class DataDisk extends AbstractModel{
      * @return DiskSize 数据盘大小，单位：GB。最小调整步长为10G，不同数据盘类型取值范围不同，具体限制详见：[CVM实例配置](https://cloud.tencent.com/document/product/213/2177)。默认值为0，表示不购买数据盘。更多限制详见产品文档。
 注意：此字段可能返回 null，表示取不到有效值。
      */
-    public Long getDiskSize() {
+    public Integer getDiskSize() {
         return this.DiskSize;
     }
 
@@ -82,7 +82,7 @@ public class DataDisk extends AbstractModel{
      * @param DiskSize 数据盘大小，单位：GB。最小调整步长为10G，不同数据盘类型取值范围不同，具体限制详见：[CVM实例配置](https://cloud.tencent.com/document/product/213/2177)。默认值为0，表示不购买数据盘。更多限制详见产品文档。
 注意：此字段可能返回 null，表示取不到有效值。
      */
-    public void setDiskSize(Long DiskSize) {
+    public void setDiskSize(Integer DiskSize) {
         this.DiskSize = DiskSize;
     }
